+++ plugin/main.ts
@@ -21,7 +21,7 @@ import {
 	shareOnlyEdited
 } from "./commands";
 import {commands, StringFunc, t, translationLanguage} from "./i18n";
-import {getTitleField, regexOnFileName} from "./contents_conversion/filePathConvertor";
+import {getTitleField, regexOnFileName} from "./conversion/filePathConvertor";
 
 /**
  * Main class of the plugin
